Move extensions to bindings/extensions.rs
